WIP: checking in so we can discuss in Tuesday meeting. Not currently working! In this commit: restructuring of code into subdirs, refactoring of build() function, initial debugging of dot-overloading approach for components.
