+++ foo.jl
@@ -1,12 +1,10 @@
-using Mimi
-
 module foo2
 using Mimi
 
 @deftimestep bar function run(v,p,i,t)
     v.a = p.b[3]
 
-    v.a[4] = p.c
+    v.a[4,2] = p.c
 end
 
 end
