+++ src/modelinstance/deftimestep_macro.jl
@@ -1,23 +1,39 @@
-function replace_dots(e)
-    if isa(e, Expr) && e.head == Symbol("=") && e.args[1].head == Symbol(".")
-        return :( set_property!( $(e.args[1].args[1]), Val($(e.args[1].args[2])), $(e.args[2]) ) )
-    elseif isa(e, Expr) && e.head == Symbol(".")
-        return :( get_property!( $(e.args[1]), Val($(e.args[2])) ) )
+#
+# Main> MacroTools.prewalk(replace_dots, :(p.foo[1,2] = v.bar))
+# :((get_property(p, Val(:foo)))[1, 2] = get_property(v, Val(:bar)))
+#
+# Main> MacroTools.prewalk(replace_dots, :(p.foo = v.bar[1]))
+# :(set_property!(p, Val(:foo), (get_property(v, Val(:bar)))[1]))
+#
+function replace_dots(ex)
+    if @capture(ex, obj_.field_ = rhs_)
+        return :(set_property!($obj, Val($(QuoteNode(field))), $rhs))
+    
+    elseif @capture(ex, obj_.field_)
+        return :(get_property($obj, Val($(QuoteNode(field)))))
+
+    elseif @capture(ex, obj_.field_[args__] = rhs_)
+        return :(get_property($obj, Val($(QuoteNode(field))))[$(args...)] = $rhs)
+
+    elseif @capture(ex, obj_.field_[args__])
+        return :(get_property($obj, Val($(QuoteNode(field))))[$(args...)])
+
     else
-        return e
+        return ex
     end
 end
 
-# This is a temporary macro until we move the run_timestep definition
-# into the @defcomp macro. For now this macro can be put before a
-# `function run_timestep` definition.
 macro deftimestep(compname, ex)
     module_name = Symbol(string(current_module()))
-    q = quote
-        function run_timestep(::Val{$(QuoteNode(module_name))}, ::Val{$(QuoteNode(compname))}, $(ex.args[1].args[2]),$(ex.args[1].args[3]),$(ex.args[1].args[4]),$(ex.args[1].args[5]))
-            $(MacroTools.prewalk(replace_dots,ex. args[2]))
-        end
+    if ! @capture(ex, function run(args__) body__ end)
+        error("Badly formatted timestep function: $ex")
     end
-    # println(q)
-    return q
+
+    func = :(
+        function run_timestep(::Val{$(QuoteNode(module_name))}, ::Val{$(QuoteNode(compname))}, $(args...))
+            $(MacroTools.prewalk(replace_dots, body)...)
+        end
+    )
+    # println(func)
+    return func
 end
