+++ src/modelinstance/dotoverloading.jl
@@ -1,4 +1,5 @@
-@generated function get_property(v::ModelInstanceComponentVariables{NAMES,TYPES}, ::Val{PROPERTYNAME}) where {NAMES,TYPES,PROPERTYNAME}
+@generated function get_property(v::ModelInstanceComponentVariables{NAMES,TYPES}, 
+                                 ::Val{PROPERTYNAME}) where {NAMES,TYPES,PROPERTYNAME}
     index_pos = findfirst(NAMES, PROPERTYNAME)
     index_pos==0 && error("Unknown variable name $PROPERTYNAME.")
 
@@ -9,7 +10,8 @@
     end
 end
 
-@generated function set_property(v::ModelInstanceComponentVariables{NAMES,TYPES}, ::Val{PROPERTYNAME}, value) where {NAMES,TYPES,PROPERTYNAME}
+@generated function set_property!(v::ModelInstanceComponentVariables{NAMES,TYPES}, 
+                                  ::Val{PROPERTYNAME}, value) where {NAMES,TYPES,PROPERTYNAME}
     index_pos = findfirst(NAMES, PROPERTYNAME)
     index_pos==0 && error("Unknown variable name $PROPERTYNAME.")
 
@@ -20,7 +22,8 @@ end
     end
 end
 
-@generated function get_property(v::ModelInstanceComponentParameters{NAMES,TYPES}, ::Val{PROPERTYNAME}) where {NAMES,TYPES,PROPERTYNAME}
+@generated function get_property(v::ModelInstanceComponentParameters{NAMES,TYPES}, 
+                                 ::Val{PROPERTYNAME}) where {NAMES,TYPES,PROPERTYNAME}
     index_pos = findfirst(NAMES, PROPERTYNAME)
     index_pos==0 && error("Unknown variable name $PROPERTYNAME.")
