+++ src/utils/graph.jl
@@ -1,0 +1,54 @@
+#
+# Graph Functionality
+#
+function show(io::IO, m::Model)
+    println(io, "showing model component connections:")
+    for item in enumerate(keys(m.components2))
+        c = item[2]
+        i_connections = get_connections(m,c,:incoming)
+        o_connections = get_connections(m,c,:outgoing)
+        println(io, item[1], ". ", c, " component")
+        println(io, "    incoming parameters:")
+        if length(i_connections)==0
+            println(io, "      none")
+        else
+            [println(io, "      - ",e.target_parameter_name," from ",e.source_component_name," component") for e in i_connections]
+        end
+        println(io, "    outgoing variables:")
+        if length(o_connections)==0
+            println(io, "      none")
+        else
+            [println(io, "      - ",e.source_variable_name," in ",e.target_component_name, " component") for e in o_connections]
+        end
+    end
+end
+
+function get_connections(m::Model, c::ComponentInstanceInfo, which::Symbol)
+    return get_connections(m, c.name, which)
+end
+
+function get_connections(m::Model, component_name::Symbol, which::Symbol)
+    if which==:all
+        f = e -> e.source_component_name==component_name || e.target_component_name==component_name
+    elseif which==:incoming
+        f = e -> e.target_component_name==component_name
+    elseif which==:outgoing
+        f = e -> e.source_component_name==component_name
+    else
+        error("Invalid parameter for the 'which' argument; must be 'all' or 'incoming' or 'outgoing'.")
+    end
+    return collect(Iterators.filter(f, m.internal_parameter_connections))
+end
+
+function get_connections(mi::ModelInstance, component_name::Symbol, which::Symbol)
+    if which==:all
+        f = e -> e.source_component_name==component_name || e.target_component_name==component_name
+    elseif which==:incoming
+        f = e -> e.target_component_name==component_name
+    elseif which==:outgoing
+        f = e -> e.source_component_name==component_name
+    else
+        error("Invalid parameter for the 'which' argument; must be 'all' or 'incoming' or 'outgoing'.")
+    end
+    return collect(Iterators.filter(f, mi.internal_parameter_connections))
+end
