+++ test/test_references.jl
@@ -1,4 +1,4 @@
-using Base.Test
+#using Base.Test
 using Mimi
 
 @defcomp Foo begin
@@ -26,8 +26,8 @@ bar = addcomponent(m, Bar)
 
 foo[:input] = 3.14
 bar[:intermed] = foo[:intermed]
-#connectparameter(m, :Bar, :intermed, :Foo, :intermed)
+# connectparameter(m, :Bar, :intermed, :Foo, :intermed)
 
 run(m)
 
-@test m[:Bar, :output][1] == 3.14
+# @test m[:Bar, :output][1] == 3.14
